Added Dijkstra and Bellman-Ford

+++ src/main/java/graphs/Edge.java
@@ -3,9 +3,9 @@ package graphs;
 public class Edge {
     public String from;
     public String to;
-    public int weight;
+    public double weight;
 
-    public Edge(String from, String to, int weight) {
+    public Edge(String from, String to, double weight) {
         this.from = from;
         this.to = to;
         this.weight = weight;
